# ChartTabs: Optimize performance using the data layer