--- a/client/components/data/query-site-stats/constants.js
+++ b/client/components/data/query-site-stats/constants.js
@@ -1,0 +1,3 @@
+/** @format */
+
+export const DEFAULT_HEARTBEAT = 30 * 60 * 1000; // 30 minutes