--- a/client/state/data-layer/wpcom/sites/stats/index.js
+++ b/client/state/data-layer/wpcom/sites/stats/index.js
@@ -5,6 +5,7 @@
  */
 import { mergeHandlers } from 'state/action-watchers/utils';
 import recentPostViews from './views/posts';
+import chartCounts from './visits';
 import googleMyBusiness from './google-my-business';
 
-export default mergeHandlers( recentPostViews, googleMyBusiness );
+export default mergeHandlers( recentPostViews, googleMyBusiness, chartCounts );