--- a/client/state/data-layer/wpcom/sites/stats/visits/from-api.js
+++ b/client/state/data-layer/wpcom/sites/stats/visits/from-api.js
@@ -1,0 +1,10 @@
+/** @format */
+
+/**
+ * Internal dependencies
+ */
+import makeJsonSchemaParser from 'lib/make-json-schema-parser';
+import { parseChartData } from 'state/stats/lists/utils';
+import responseSchema from './schema';
+
+export default makeJsonSchemaParser( responseSchema, parseChartData );