--- a/client/state/data-layer/wpcom/sites/stats/visits/index.js
+++ b/client/state/data-layer/wpcom/sites/stats/visits/index.js
@@ -1,4 +1,8 @@
 /** @format */
+/**
+ * External Dependencies
+ */
+import { difference } from 'lodash';
 
 /**
  * Internal Dependencies
@@ -11,9 +15,11 @@ import { registerHandlers } from 'state/data-layer/handler-registry';
 import fromApi from './from-api';
 
 export const fetch = action => {
-	const { date, period, quantity, siteId, statFields } = action;
+	const { chartTab, date, period, quantity, siteId, statFields } = action;
+	const currentTabFields = chartTab === 'views' ? [ 'views', 'visitors' ] : [ chartTab ];
+	const otherTabFields = difference( statFields, currentTabFields );
 
-	return statFields.map( statField =>
+	return [
 		http(
 			{
 				method: 'GET',
@@ -23,12 +29,26 @@ export const fetch = action => {
 					unit: period,
 					date,
 					quantity,
-					stat_fields: statField,
+					stat_fields: currentTabFields.join( ',' ),
 				},
 			},
 			action
-		)
-	);
+		),
+		http(
+			{
+				method: 'GET',
+				path: `/sites/${ siteId }/stats/visits`,
+				apiVersion: '1.1',
+				query: {
+					unit: period,
+					date,
+					quantity,
+					stat_fields: otherTabFields.join( ',' ),
+				},
+			},
+			action
+		),
+	];
 };
 
 export const onSuccess = ( { siteId, period }, data ) => receiveChartCounts( siteId, period, data );