--- a/client/state/data-layer/wpcom/sites/stats/visits/schema.js
+++ b/client/state/data-layer/wpcom/sites/stats/visits/schema.js
@@ -1,0 +1,25 @@
+/** @format */
+
+export default {
+	id: 'sites-stats-visits-response',
+	title: 'Stats Visits Response',
+	type: 'object',
+	$schema: 'http://json-schema.org/draft-04/schema#',
+	properties: {
+		data: {
+			type: 'array',
+			items: {
+				type: 'array',
+				items: {
+					type: [ 'integer', 'string', 'array' ],
+				},
+			},
+		},
+		fields: {
+			type: 'array',
+		},
+		unit: {
+			type: 'string',
+		},
+	},
+};