--- a/client/state/stats/chart-tabs/constants.js
+++ b/client/state/stats/chart-tabs/constants.js
@@ -1,0 +1,3 @@
+/** @format */
+
+export const QUERY_FIELDS = [ 'views', 'visitors', 'likes', 'comments', 'post_titles' ];