… client/state/action-types.js
… client/state/data-layer/wpcom/sites/stats/visits/test/index.js
… client/state/stats/chart-tabs/selectors.js
… client/state/stats/chart-tabs/test/reducer.js
… client/state/stats/chart-tabs/test/selectors.js
… client/state/stats/reducer.js