@jmdodd 2018-10-25
@jsnmoon How did you measure and compare performance with parallelized network requests versus with the original two API `stats/visits` requests?