@jsnmoon 2018-10-25
@jmdodd: Following up from our discussion on Slack, I've removed the parallelized network requests from this PR and have restored the original request strategy. 

I've also introduced an optimization to the original request strategy to reduce redundancy. Let's say the user has the `Views` tab selected:
- Before this tweak, we'd issue a second request for all fields.
  1. A request for `views,visitors` 
  2. A request for `views,visitors,likes,comments,post_titles`
- After tweak, we issue a second request for only the fields we need.
  1. A request for `views,visitors` 
  2. A request for `likes,comments,post_titles`

This optimization should alleviate the load on the API.