@sgomes 2018-10-26
Thanks for the testing instructions, @jsnmoon ! Are we specifically looking for calls to the `stats` endpoint here? If so, I'm only seeing one connection here, not two.

I'll come back to this PR on Monday with some performance analysis in terms of before/after. Hopefully We'll be able to quantize things a bit :)