@jsnmoon 2018-10-26
> Thanks for the testing instructions, @jsnmoon ! Are we specifically looking for calls to the `stats` endpoint here? If so, I'm only seeing one connection here, not two.

@sgomes: There ought to be two requests to `/stats/visits`, like so:

<img width="1284" alt="Network requests from current production" src="https://user-images.githubusercontent.com/4044428/47585008-65323880-d919-11e8-8328-f35d50d440a5.png">

I updated the original post to clarify. Hope this helps!