@sgomes 2018-10-29
Hi @jsnmoon !

I ran a bunch of tests today, and I couldn't find any evidence of performance improvements or any change in terms of number of calls to `/stats/visits`. I rolled back to 3dab7389a87dc574f472521fdfd5d8697402542d and at that commit there seem to be only two `visits` requests made as well, and with the same query string:

![image](https://user-images.githubusercontent.com/409615/47650374-f56db900-db77-11e8-8179-a0c50aeb316c.png)

I measured the performance using `perfmon` under a simulated 3G connection when loading the stats page from scratch (that is, navigating to it), when opening it for the first time (having navigated to a different page first), and when making repeated visits to it, and I couldn't find a meaningful performance delta in any of those scenarios.

Did I miss anything or do anything wrong? Is there perhaps a situation in which the benefits are more evident?