@jsnmoon 2018-10-29
Hi @sgomes,

It's strange that the second request in your screenshot includes `visitors` in the query string. Is your setup running the latest commit (94a12bed4e787817e106ead692b5d2b61c3def4d)? 

I was able to reproduce the different queries in [this Calypso live branch](https://hash-94a12bed4e787817e106ead692b5d2b61c3def4d.calypso.live/). (For easy reference, here's [master running on a live branch](https://hash-564fd7b47a3fdc82166f7ac861b47f1b59ae2a37.calypso.live/).

<img width="674" alt="screen shot 2018-10-29 at 10 42 22 am" src="https://user-images.githubusercontent.com/4044428/47665719-6193f100-db67-11e8-8fdb-92cb54381964.png">

Note that `views,visitors` has been excluded from the second request in the screenshot above.