@jsnmoon 2018-11-05
I ran some performance profiling at an internal P2 using Calypso.live at two hashes: 94a12bed4e787817e106ead692b5d2b61c3def4d and 3dab7389a87dc574f472521fdfd5d8697402542d. 

---

# Day (Avg of 5 requests)
## Without change
- Visitor request: 207.4 ms
- Supplemental request: 6.428s

## With change
- Visitor request: 170.2ms
- Supplemental request: 6.01s

## Delta
- Visitor request: -37.2ms
- Supplemental request: -418ms

# Week  (Avg of 10 requests)
## Without change
- Visitor request: 243.0ms
- Supplemental request: 7.54s

## With change
- Visitor request: 261.5ms
- Supplemental request: 7.91s

## Delta
- Visitor request: +18.5ms 
- Supplemental request: +370ms

# Month  (Avg of 10 requests)
## Without change
- Visitor request: 596.8ms
- Supplemental request: 9.597s

## With change
- Visitor request: 603.1ms
- Supplemental request: 8.93s

## Delta
- Visitor request: +6.3ms
- Supplemental request: -667ms

---

Takeaway: There might be an improvement in network performance, but it'd be hasty to make this conclusion given that there's a lot of variance between each network request (not to mention Week requests actually did worse). However, I believe we should push out this change since it removes [requesting redundant data fields](https://github.com/Automattic/wp-calypso/pull/27997/files#diff-b61543a6dc80031e3a9fcb3d8e280f5cR20) from the `stats/visits` API route.