@sgomes 2018-11-06
Thanks, @jsnmoon ! That's great.

I agree that even if the performance numbers are inconclusive this PR should be merged, as the refactoring merits are sufficient on their own. 👍